Characters print error fix in windows console
Damn it! Character cannot be out printed normally in my windows console.
I considered it as the cause of python encoding environment.
Eventually, the root cause is the code page of my windows, it was 437 as
default due to my US locale. Now I change the locale to Chinese so that
it works well.